requirements.txt
python-dotenv
httpx
pillow
requests

mcp-core
mcp-server
mcp-server-fetch
mcp
exa-py

pyautogen
autogen-ext

fastapi
uvicorn
pydantic
openai
autogen-agentchat
autogen-core
torch
transformers